requirements.txt
Flask==3.0.0
pandas==2.0.3
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
